chore(etl): IDE 类型噪音清零 —— bs4/openpyxl 存根入 dev 组,六处 find 收窄,int(round) 摘冗
Frank 逐黄点排查:employers「2」= int(round()) 多余转换(py3 单参 round 本就返 int,
三处摘冗值不变);lmia 黄点 = openpyxl 无存根(types-openpyxl/types-beautifulsoup4
入 dev 组)。bs4 真存根落地后 find() 的 Tag|NavigableString|None 真形现身,crawl×3 /
fetch×3 的裸用改 isinstance(Tag) 收窄(实际数据恒 Tag,行为等价;诡异页面从崩改跳)。
pyrefly 0 错、警告 10→5(余为无存根库的 untyped-import,官方未发存根,不可行动)。

[skip render]

+++ etl/employers/functions.py
@@ -59,7 +59,7 @@ def median_of(values: list) -> int | None:
     if len(vals) == 0:
         return None
     vals.sort()
-    return int(round(vals[len(vals) // 2]))
+    return round(vals[len(vals) // 2])
 
 
 def skilled_broads_of(lmia_row: dict) -> dict:
@@ -327,13 +327,13 @@ def bucket_row_of(x: BucketIn) -> BucketRow:
     scan = bucket_scan_of(rows)
     share = None
     if len(rows) > 0:
-        share = int(round(WAGE_INDEX_BASE * scan.entry / len(rows)))
+        share = round(WAGE_INDEX_BASE * scan.entry / len(rows))
     wage_med = median_of(scan.wages)
     index = None
     if wage_med is not None and scan.prov_top:
         cell_med = median_of(ctx.wage_cells.get((x.broad, scan.prov_top)) or [])
         if cell_med:
-            index = int(round(WAGE_INDEX_BASE * wage_med / cell_med))
+            index = round(WAGE_INDEX_BASE * wage_med / cell_med)
     lmia_n = skilled_broads_of(lmia_row).get(x.broad, 0)
     star = star_of(StarIn(designated=len(des_rows) > 0, open_jobs=len(rows),
                           entry_jobs=scan.entry, lmia_skilled=lmia_n))
